Skin_Tumor_Classification_CNN
This is a first attempt of using the pre-trained neural network ResNet50 to classify skin tumor imagines in Matlab.      Note: this is the first draft, and for rapidity purposes it was trained for on a small data base. The final version will be applied on a much larger image set.